Adjust ch.getting_old formula.
We want to fix the overflow bug in future.
But before that, we should make the whole getting old calculation more feasible
since most people can't access their pets every 6 hours. (Or, we can say,
the chicken now is different from how it was invented before - it's now a more
long-term game for most people)