Allow input prefixes that matches no user in usercomplete.
It wasn't possible to input any char that makes a completion matches
nothing. For example, if there's no userid that starts with "foo", it's
not possible to type "foo" into the auto-usercompletion input.

This commit makes the former mentioned input possible. It has some small
benefits such as less cpu usage and memory allocation, due to less
vector construction. Though it might be a bit controversial whether this
kind of input should be allowed.

The intention of this commit is to make supporting remote user db
storage easier. We only need to do a prefix query when user hits space
or enter under this setting.