Remove OUTTA_TIMER mechanism
In the early days, OUTTA_TIMER provided a syscall-less mechanism to get
the approximate current time. This was implemented with a time value
stored in shared memory, which is updated every second by a separate
daemon.

On modern Linux and FreeBSD systems, the kernel provides a fast syscall
gettimeofday implementation that does not actually do a system call into
the kernel. Named vDSO on Linux, this is done by the kernel dynamically
mapping shared memory and a small code segment for gettimeofday (and
some other syscalls). gettimeofday retrieves the current time value from
the shared memory segment. This value is directly updated by the kernel
on each timer tick. In a sense this is a system-wide OUTTA_TIMER.

Thus we can get rid of our own custom implementation.