Install daemon binaries with full date suffix
Over the years we've accumulated a bunch of old binaries on our systems.
The old file naming scheme did not include the year of when the binary
was installed. This leads to a bit of confusion when trying to locate a
certain version.

Install the binaries with the year included.