mbbsd: Declare helper functions as static instead of inline
inline without a static confuses GCC whether to inline or not. The
result is sometimes it is not inlined, but the helper function is
not compiled as a separate function either. This results in the
linker not being able to find the function.

Keeping with best practice, mark them as static instead, and let the
compiler decide whether to inline or not.