regmaild: Add verifydb that stores extra user verification info
New table `verifydb` storing
(userid, generation, vmethod, vkey, timestamp).

Primary key (userid, generation, vmethod) tuple is used to uniquely
identify an user with an account verification method in the life time of
system, where generation is the firstlogin timestamp in the userec_t.
userid is stored in lowercase.

timestamp field is informational storing when the vkey is set.

(vmethod, vkey) uniquely indentifies a account verificaion. vkey should
be the unique subkey in vmethod. Several important notes:
* vkey should be generated in an uniform way, for instance, in
  lowercase. regmaildb does not transform vkey in API call.
* (vmethod, vkey) is indexed for fast lookup.
* There is no constaint set on the index that avoids duplication.

This is a framework and does not define any vmethod yet. Here is an
example of possible record in the future:
("sysop", 123, VMETHOD_EMAIL, "sysop@ptt.cc", 456)