regmaild: Add init db for verifydb.
initemaildb now takes a new argument on which table to initialize.

verifydb is initialized with users that have
* completed registration (PERM_LOGINOK), and
* valid email.

The timestamp is set to 0 for each record as there is no good way to
determine it from userec_t.

(vmethod, vkey) is in (VMETHOD_EMAIL, "sysop@ptt.cc") format, email
transformed to lowercase.