mbbsd: register_sms: Fix misleading default for user agreement
The user agreement requires an explicit "yes" (agree) from the user.
However the prompt is misleading in that 'Y" is capitalized, implying
that it is the default. This leads to users just pressing enter, and
returning to the menu.

Change 'Y" to lower case to signal that an explicit 'y' is required.