mbbsd: remind setting contact-email after successfully login (in user_login).
mbbsd: add check_contact_email
util: add migrate_is_contact_email

Currently the contact-email is mixed with the register-email, letting us unable
to know whether the email is contact-email or just the register-email.

Adding 1-byte-flag (uint8_t) is_contact_email in userec_t to know whether the email
is contact-email or not.

Setup the flag as long as we change the contact-email (even if the email is the same).

Previously we don't set is_contact_email in userec_t, but we have the logs in user.security.
Use user.security to migrate is_contact_email.

To deploy this update:
1. make all (no install)
2. do migrate_is_contact_email from /home/bbs
3. make install

To discuss:
whether we should increase the SHM version?
considering that there is no real break change with the original version.