mbbsd: add email_challenge
Currently we verify email again only in AccountRecovery.
We should verify email as well when the user intends to reset the password.

It is possible that we will need to verify the email when changing
the contact-email in the future.
Use email_challenge instead of passwd_challenge for this generalization.

The implementation is based on ::AccountRecovery::EmailCodeChallenge.

::AccountRecovery::EmailCodeChallenge is integrated with emailchallenge::EmailChallenge.