pfterm: fix broken scrolling when term height > ft.rows
fterm_rawscroll() required that term height == ft.rows in order to work properly,
but it is not the case when ft.rows is a clipped value of term height.

fterm_rawscroll() now only assumes the term height to be <= 9999.

Since the actual term height might not be available to pfterm,
9999 instead of a dedicated variable is used to keep existing code work.